+++ PacklinkPro/Bootstrap.php
@@ -2,11 +2,14 @@
 /**
  * @package    Packlink_PacklinkPro
  * @author     Packlink Shipping S.L.
- * @copyright  2020 Packlink
+ * @copyright  2021 Packlink
  */
 
 namespace Packlink\PacklinkPro;
 
+use Magento\Backend\Model\Auth\Session;
+use Magento\Store\Model\Information;
+use Magento\Store\Model\StoreManagerInterface;
 use Packlink\PacklinkPro\Entity\QuoteCarrierDropOffMapping;
 use Packlink\PacklinkPro\IntegrationCore\BusinessLogic\BootstrapComponent;
 use Packlink\PacklinkPro\IntegrationCore\BusinessLogic\Configuration;
@@ -36,6 +39,8 @@ use Packlink\PacklinkPro\Services\BusinessLogic\ShopOrderService;
 use Packlink\PacklinkPro\Services\BusinessLogic\UserAccountService;
 use Packlink\PacklinkPro\IntegrationCore\BusinessLogic\User\UserAccountService as BaseUserAccountService;
 use Packlink\PacklinkPro\Services\Infrastructure\LoggerService;
+use Packlink\PacklinkPro\IntegrationCore\BusinessLogic\Registration\RegistrationInfoService as RegistrationInfoServiceInterface;
+use Packlink\PacklinkPro\Services\BusinessLogic\RegistrationInfoService;
 
 /**
  * Class Bootstrap
@@ -74,6 +79,18 @@ class Bootstrap extends BootstrapComponent
      * @var UserAccountService
      */
     private $userAccountService;
+    /**
+     * @var Session
+     */
+    private $session;
+    /**
+     * @var Information
+     */
+    private $information;
+    /**
+     * @var StoreManagerInterface
+     */
+    private $storeManager;
 
     /**
      * Bootstrap constructor.
@@ -84,6 +101,9 @@ class Bootstrap extends BootstrapComponent
      * @param ShopOrderService $shopOrderService
      * @param CarrierService $carrierService
      * @param UserAccountService $userAccountService
+     * @param \Magento\Backend\Model\Auth\Session $session
+     * @param \Magento\Store\Model\Information $information
+     * @param \Magento\Store\Model\StoreManagerInterface $storeManager
      */
     public function __construct(
         CurlHttpClient $httpClientService,
@@ -91,7 +111,10 @@ class Bootstrap extends BootstrapComponent
         ConfigurationService $configService,
         ShopOrderService $shopOrderService,
         CarrierService $carrierService,
-        UserAccountService $userAccountService
+        UserAccountService $userAccountService,
+        Session $session,
+        Information $information,
+        StoreManagerInterface $storeManager
     ) {
         $this->httpClientService = $httpClientService;
         $this->loggerService = $loggerService;
@@ -99,6 +122,9 @@ class Bootstrap extends BootstrapComponent
         $this->shopOrderService = $shopOrderService;
         $this->carrierService = $carrierService;
         $this->userAccountService = $userAccountService;
+        $this->session = $session;
+        $this->information = $information;
+        $this->storeManager = $storeManager;
 
         static::$instance = $this;
     }
@@ -195,6 +221,17 @@ class Bootstrap extends BootstrapComponent
             BaseUserAccountService::CLASS_NAME,
             function () use ($instance) {
                 return $instance->userAccountService;
+            }
+        );
+
+        ServiceRegister::registerService(
+            RegistrationInfoServiceInterface::CLASS_NAME,
+            function () {
+                return new RegistrationInfoService(
+                    $this->session,
+                    $this->information,
+                    $this->storeManager
+                );
             }
         );
     }