+++ PacklinkPro/Controller/Adminhtml/Configuration/Registration.php
@@ -2,7 +2,7 @@
 /**
  * @package    Packlink_PacklinkPro
  * @author     Packlink Shipping S.L.
- * @copyright  2020 Packlink
+ * @copyright  2021 Packlink
  */
 
 namespace Packlink\PacklinkPro\Controller\Adminhtml\Configuration;
@@ -10,8 +10,7 @@ namespace Packlink\PacklinkPro\Controller\Adminhtml\Configuration;
 use Magento\Backend\App\Action\Context;
 use Magento\Framework\Controller\Result\JsonFactory;
 use Packlink\PacklinkPro\Bootstrap;
-use Packlink\PacklinkPro\IntegrationCore\BusinessLogic\Country\CountryService;
-use Packlink\PacklinkPro\IntegrationCore\Infrastructure\ServiceRegister;
+use Packlink\PacklinkPro\IntegrationCore\BusinessLogic\Controllers\RegistrationController;
 
 /**
  * Class Registration
@@ -20,6 +19,18 @@ use Packlink\PacklinkPro\IntegrationCore\Infrastructure\ServiceRegister;
  */
 class Registration extends Configuration
 {
+    /**
+     * Array that identifies e-commerce.
+     *
+     * @var string[]
+     */
+    protected static $ecommerceIdentifiers = ['Magento'];
+
+    /**
+     * @var RegistrationController
+     */
+    private $baseController;
+
     /**
      * Registration constructor.
      *
@@ -35,21 +46,51 @@ class Registration extends Configuration
         parent::__construct($context, $bootstrap, $jsonFactory);
 
         $this->allowedActions = [
-            'getSupportedCountries',
+            'getRegisterData',
+            'register'
         ];
+
+        $this->baseController = new RegistrationController();
     }
 
-    protected function getSupportedCountries()
+    /**
+     * Returns registration data.
+     *
+     * @return \Magento\Framework\Controller\Result\Json
+     */
+    protected function getRegisterData()
     {
-        /** @var CountryService $countryService */
-        $countryService = ServiceRegister::getService(CountryService::CLASS_NAME);
-        $supportedCountries = $countryService->getSupportedCountries();
+        $request = $this->getRequest();
 
-        foreach ($supportedCountries as $country) {
-            $country->registrationLink = str_replace('magento', 'pro', $country->registrationLink);
-            $country->name = __($country->name);
+        if (!$request->getParam('country')) {
+            return $this->formatNotFoundResponse();
         }
 
-        return $this->formatDtoEntitiesResponse($supportedCountries);
+        return $this->result->setData($this->baseController->getRegisterData($request->getParam('country')));
+    }
+
+    /**
+     * Attempts to register the user on Packlink PRO.
+     *
+     * @return \Magento\Framework\Controller\Result\Json
+     */
+    protected function register()
+    {
+        $payload = $this->getPacklinkPostData();
+
+        $payload['ecommerces'] = static::$ecommerceIdentifiers;
+
+        try {
+            $status = $this->baseController->register($payload);
+            return $this->result->setData(['success' => $status]);
+        } catch (\Exception $e) {
+            return $this->result->setData(
+                [
+                    'success' => false,
+                    'error' => $e->getMessage() === 'Registration failed. Error: ' ?
+                        'Registration failed. Error: Invalid phone number.' : $e->getMessage(),
+                ]
+            );
+        }
     }
 }