+++ deploy.sh
@@ -35,7 +35,13 @@ removeUnnecessaryFiles() {
   rm -rf packlink/composer.lock
   rm -rf packlink/PacklinkPro/Test
   rm -rf packlink/PacklinkPro/IntegrationCore/Tests
+  rm -rf packlink/PacklinkPro/IntegrationCore/DemoUI
+  rm -rf packlink/PacklinkPro/IntegrationCore/BusinessLogic/Resources/lang/translations.csv
+  rm -rf packlink/PacklinkPro/IntegrationCore/BusinessLogic/Resources/lang/fromCSV.php
+  rm -rf packlink/PacklinkPro/IntegrationCore/BusinessLogic/Resources/lang/toCSV.php
   rm -rf packlink/PacklinkPro/IntegrationCore/Infrastructure/Serializer/Concrete/NativeSerializer.php
+  rm -rf packlink/PacklinkPro/view/adminhtml/web/packlink/lang/fromCSV.php
+  rm -rf packlink/PacklinkPro/view/adminhtml/web/packlink/lang/toCSV.php
 }
 
 createZipArchives() {